refactor(matrix): share boolean param parsing utility
